Use SSO and allow multiple users

+++ server.js
@@ -1,9 +1,10 @@
 const express = require("express")
 const app = express()
-const redirectToHTTPS = require("express-http-to-https").redirectToHTTPS
-const cookieParser = require("cookie-parser")
-const hash = require("password-hash")
+const session = require("express-session")
+const redisStore = require("connect-redis")(session)
+const redis = require("ioredis")
 const fs = require("fs")
+const axios = require("axios")
 
 const database = require("./db.js")
 const countryDropdown = require("./countryDropdown.js")
@@ -14,26 +15,41 @@ const html = fs.readFileSync(__dirname + "/views/index.html").toString().replace
 app.use(express.static(__dirname + "/public"))
 app.use(express.urlencoded({extended: false}))
 app.use(express.json())
-app.use(redirectToHTTPS([/localhost:(\d{4})/]))
-app.use(cookieParser())
+app.set("trust proxy", 1)
 
-app.post('/login', (req, res) => {
-  // Check the password
-  console.log(req.body.password)
-  if (hash.verify(req.body.password, process.env.PASSWORD)) {
-    res.cookie("token", process.env.ACCESS_TOKEN, {maxAge: 2147483647})
+let redisClient = new redis(process.env.REDIS_URL)
+app.use(session({
+  store: new redisStore({client: redisClient}),
+  secret: "something random",
+  saveUninitialized: false,
+  resave: false
+}))
+
+// Callback url from SSO
+app.get("/sso", (req, res) => {
+  let url = new URL("https://auth.alexclifton.co.uk/verifyToken")
+  url.searchParams.append("token", req.query.token)
+  url.searchParams.append("apikey", process.env.SSO_KEY)
+  axios.get(url.href).then(response => {
+    if (response.data.error) {
+      res.send("<h1>Auth error</h1><br>" + response.data.error)
+      return
+    }
+    req.session.user = {
+      id: response.data.id,
+      name: response.data.name,
+      admin: response.data.admin == "true"
+    }
     res.redirect("/")
-  } else {
-    res.send("Incorrect Password")
-  }
+  })
 })
 
-app.all('*', (req, res, next) => {
-  // Check if token is valid
-  if (req.cookies.token == process.env.ACCESS_TOKEN) {
+// On all requests, verify user is logged in
+app.all("*", (req, res, next) => {
+  if (req.session.user) {
     next()
   } else {
-    res.sendFile(__dirname + "/views/login.html")
+    res.redirect("https://auth.alexclifton.co.uk/login?app=" + process.env.SSO_APP)
   }
 })
 
@@ -43,9 +59,9 @@ app.get("/", (req, res) => {
 
 // Send all diary entries
 app.get("/diary", (req, res) => {
-  let sql = "SELECT rowid, date, country, place, notes FROM entries ORDER BY date DESC, country ASC, place ASC"
+  let sql = "SELECT rowid, date, country, place, notes FROM entries WHERE public = true OR owner = $1 ORDER BY date DESC, country ASC, place ASC"
   let db = database.connect()
-  db.query(sql, (err, data) => {
+  db.query(sql, [req.session.user.id],(err, data) => {
     if (err) throw err
     res.send(data.rows)
     db.end()
@@ -54,20 +70,24 @@ app.get("/diary", (req, res) => {
 
 // Send a single entry
 app.get("/entry", (req, res) => {
-  let sql = "SELECT date, country, place, notes FROM entries WHERE rowid = $1"
+  let sql = "SELECT date, country, place, public, notes, owner FROM entries WHERE rowid = $1 AND (public = true OR owner = $2)"
   let db = database.connect()
-  db.query(sql, [req.query.id], (err, data) => {
+  db.query(sql, [req.query.id, req.session.user.id], (err, data) => {
     if (err) throw err
-    res.send(data.rows[0])
+    
+    let entry = data.rows[0]
+    // Work out if the user can edit this entry
+    entry.canEdit = (req.session.user.id == entry.owner) || req.session.user.admin
+    res.send(entry)
     db.end()
   })
 })
 
 // Send unique values for filters
 app.get("/filterValues", (req, res) => {
-  let sql = "SELECT DISTINCT EXTRACT(year FROM to_timestamp( CAST( date AS bigint ) / 1000 )) AS year FROM entries ORDER BY year;"
+  let sql = "SELECT DISTINCT EXTRACT(year FROM to_timestamp( CAST( date AS bigint ) / 1000 )) AS year FROM entries WHERE public = true OR owner = $1 ORDER BY year;"
   let db = database.connect()
-  db.query(sql, (err, data) => {
+  db.query(sql, [req.session.user.id], (err, data) => {
     if (err) throw err
     let filters = {}
     filters.year = data.rows.map(x => x.year)
@@ -81,9 +101,9 @@ app.get("/filterValues", (req, res) => {
 app.post("/new", (req, res) => {
   let date = new Date(req.body.date).getTime()
   
-  let sql = "INSERT INTO entries (date, country, place, notes) VALUES ($1, $2, $3, $4)"
+  let sql = "INSERT INTO entries (date, country, place, public, notes, owner) VALUES ($1, $2, $3, $4, $5, $6)"
   let db = database.connect()
-  db.query(sql, [date, req.body.country, req.body.place, req.body.notes], (err) => {
+  db.query(sql, [date, req.body.country, req.body.place, req.body.public, req.body.notes, req.session.user.id], (err) => {
     if (err) throw err
     res.send("ok")
     db.end()
@@ -92,11 +112,14 @@ app.post("/new", (req, res) => {
 
 // Edit an entry
 app.post("/edit", (req, res) => {
-  let date = new Date(req.body.date).getTime()
+  let entry = req.body
+  let date = new Date(entry.date).getTime()
   
-  let sql = "UPDATE entries SET date = $1, country = $2, place = $3, notes = $4 WHERE rowid = $5"
+  let user = req.session.user
+  
+  let sql = "UPDATE entries SET date = $1, country = $2, place = $3, notes = $4, public = $5 WHERE rowid = $6 AND (owner = $7 OR $8)"
   let db = database.connect()
-  db.query(sql, [date, req.body.country, req.body.place, req.body.notes, req.body.id], (err) => {
+  db.query(sql, [date, entry.country, entry.place, entry.notes, entry.public, entry.id, user.id, user.admin], (err) => {
     if (err) throw err
     res.send("ok")
     db.end()
@@ -106,9 +129,11 @@ app.post("/edit", (req, res) => {
 
 // Delete an entry
 app.post("/delete", (req, res) => {
-  let sql = "DELETE FROM entries WHERE rowid = $1"
+  let user = req.session.user
+
+  let sql = "DELETE FROM entries WHERE rowid = $1 AND (owner = $2 OR $3)"
   let db = database.connect()
-  db.query(sql, [req.body.id], (err) => {
+  db.query(sql, [req.body.id, user.id, user.admin], (err) => {
     if (err) throw err
     res.send("ok")
     db.end()
